+++ src/components/manage/Blocks/Block/BlocksForm.jsx
@@ -13,6 +13,9 @@ import {
   previousBlockId,
 } from '@plone/volto/helpers';
 import EditBlockWrapper from './EditBlockWrapper';
+import { setSidebarTab } from '@plone/volto/actions';
+import { useDispatch } from 'react-redux';
+import { useDetectClickOutside } from 'react-detect-click-outside';
 import config from '@plone/volto/registry';
 
 const BlocksForm = (props) => {
@@ -35,6 +38,18 @@ const BlocksForm = (props) => {
   } = props;
 
   const blockList = getBlocks(properties);
+
+  const dispatch = useDispatch();
+
+  const ClickOutsideListener = (e) => {
+    onSelectBlock(null);
+    dispatch(setSidebarTab(0));
+  };
+
+  const ref = useDetectClickOutside({
+    onTriggered: ClickOutsideListener,
+    triggerKeys: ['Escape'],
+  });
 
   const handleKeyDown = (
     e,
@@ -129,7 +144,7 @@ const BlocksForm = (props) => {
   const editBlockWrapper = children || defaultBlockWrapper;
 
   return (
-    <div className="blocks-form">
+    <div className="blocks-form" ref={ref}>
       <DragDropList
         childList={blockList}
         onMoveItem={(result) => {